Extract core report generation logic into separate package
