+++ report/markdown.go
@@ -1,4 +1,4 @@
-package main
+package report
 
 import (
 	"fmt"
